+++ server/src/main/java/org/opensearch/indices/replication/SegmentReplicationUpgradeListener.java
@@ -21,6 +21,11 @@ import org.opensearch.indices.IndicesService;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * SegmentReplicationUpgradeListener is used to upgrade the opensearch version used by all primaries of a cluster when
+ * segment replication is enabled and a rolling upgrade is completed (while in mixed cluster state, the primaries use lower codec
+ * version on their primaries and this needs to be reset once upgrade is complete).
+ */
 public class SegmentReplicationUpgradeListener implements ClusterStateListener {
 
     private static final Logger logger = LogManager.getLogger(SegmentReplicationUpgradeListener.class);
@@ -33,29 +38,34 @@ public class SegmentReplicationUpgradeListener implements ClusterStateListener {
 
     @Override
     public void clusterChanged(ClusterChangedEvent event) {
-        List<IndexShard> indexShardList = new ArrayList<>();
-        DiscoveryNodes nodes = event.state().nodes();
-        if (nodes.getMinNodeVersion().equals(nodes.getMaxNodeVersion())) {
-            for (IndexService indexService : indicesService.indices().values()) {
-                for (IndexShard indexShard : indexService) {
-                    try {
-                        if ((indexShard.getEngine().config().isReadOnlyReplica() == false) && (indexShard.getEngine().config().getClusterMinVersion() != nodes.getMaxNodeVersion())) {
-                            indexShardList.add(indexShard);
+        if (event.nodesChanged()) {
+            List<IndexShard> indexShardList = new ArrayList<>();
+            DiscoveryNodes nodes = event.state().nodes();
+            if (nodes.getMinNodeVersion().equals(nodes.getMaxNodeVersion())) {
+                for (IndexService indexService : indicesService) {
+                    for (IndexShard indexShard : indexService) {
+                        try {
+                            if (indexShard.indexSettings().isSegRepEnabled()
+                                && indexShard.indexSettings().getNumberOfReplicas() > 0
+                                && indexShard.routingEntry().primary()
+                                && (indexShard.getEngine().config().getClusterMinVersion() != nodes.getMaxNodeVersion())) {
+                                indexShardList.add(indexShard);
+                            }
+                        } catch (AlreadyClosedException e) {
+                            logger.warn("Index shard [{}] engine is already closed.", indexShard.shardId());
                         }
-                    } catch (AlreadyClosedException e) {
-                        logger.warn("Index shard [{}] engine is already closed.", indexShard.shardId());
                     }
                 }
             }
-        }
-        try {
-            if (indexShardList.isEmpty() == false) {
-                for (IndexShard is : indexShardList) {
-                    is.resetEngineToGlobalCheckpoint();
+            try {
+                if (indexShardList.isEmpty() == false) {
+                    for (IndexShard indexShard : indexShardList) {
+                        indexShard.resetEngine();
+                    }
                 }
+            } catch (Exception e) {
+                logger.error("Received unexpected exception: [{}]", e.getMessage());
             }
-        } catch (Exception e) {
-            logger.error("Received unexpected exception: [{}]", e.getMessage());
         }
 
     }